added change location on items list
