+++ src/resources/items/ItemList.tsx
@@ -11,13 +11,20 @@ import {
   SelectInput,
   TextField,
   TextInput,
-  TopToolbar
+  TopToolbar,
+  BulkDeleteButton,
+  useListContext,
+  useRefresh
 } from 'react-admin'
 import SourceField from '../../components/SourceField'
 import SourceInput from '../../components/SourceInput'
 import { mediaTypeOptions } from '../../utils/media'
 import * as constants from '../../constants'
 import CreatedByMeFilter from '../../components/CreatedByMeFilter'
+import { Button, Modal } from '@mui/material'
+import { useState } from 'react'
+import FlexBox from '../../components/FlexBox'
+import ChangeLocation from './ItemForm/ChangeLocation'
 const sort = (field = 'name') => ({ field, order: 'ASC' })
 
 const omitColumns: string[] = ['createdAt']
@@ -67,11 +74,49 @@ const ItemActions = () => (
   </TopToolbar>
 )
 
+export const BulkActions = () => {
+  const { selectedIds } = useListContext()
+  const [open, setOpen] = useState(false)
+  const refresh = useRefresh()
+
+  const handleClose = () => {
+    setOpen(false)
+  }
+
+  const handleOpen = () => {
+    setOpen(true)
+  }
+
+  const handleSuccess = () => {
+    handleClose()
+    refresh()
+  }
+
+  return (
+    <>
+      <FlexBox>
+        <BulkDeleteButton mutationMode='pessimistic' />
+        <Button size='small' variant='outlined' onClick={handleOpen}>
+          Change Location
+        </Button>
+      </FlexBox>
+      <Modal open={open} onClose={handleClose}>
+        <ChangeLocation
+          successCallback={handleSuccess}
+          onCancel={handleClose}
+          ids={selectedIds}
+        />
+      </Modal>
+    </>
+  )
+}
+
 export default function ItemList(
   props?: Omit<ListProps, 'children'>
 ): React.ReactElement {
   return (
     <List
+      bulkActionButtons={<BulkActions />}
       hasCreate={false}
       actions={<ItemActions />}
       resource='items'
